Update module discovery to return tree instead of list
Updates the tool's Go module discovery logic so that it returns a tree
of modules instead of a list of modules. This ensure that sorting and
hierarchy of the modules are preserved.

Fixes a bug where modules that had the same root path would not
correctly be identified as sub modules, and instead get identified as
submodules of the root module.