Update release tooling to only create release tag for multi-module repos (#4)
Updates the release tooling so that calculaterelease and tagrelease only create the release-YYYY-MM-DD tag for multi-module releases. Single module repositories will use the root module's tag for the release manifest ID.

Adds a utility moduleversion to return the Go module's current version or when used with -unreleased returns the module's projected version on the next release.